KVM: arm64: Add the KVM_ARM_VCPU_SUPPORTED_CPUS VCPU ioctl
The ioctl specifies a list of CPUs which support running all the
features configured for the VCPU. The ioctl introduces no constraints on
the VCPU scheduling, and userspace is expected to manage the VCPU
affinity. Attempting to run the VCPU on a CPU not present in the list
will result in KVM_RUN returning -ENOEXEC.

The ioctl will be used by KVM to prevent errors (like accesses to
undefined registers) when emulating VCPU features that require hardware
support which is present only on a subset of the CPUs present in the
system.