drivers/perf: Expose the cpumask of CPUs that support SPE
KVM SPE emulation will require the list of CPUs that have hardware support
for SPE. Add a function that returns the cpumask of supported CPUs
discovered by the driver during probing.

The function only works if the driver is compiled with CONFIG_ARM_SPE_PMU=y
because KVM relies on the driver to enable the SPE interrupt.