KVM: arm64: Add lock/unlock memslot user API
Stage 2 faults triggered by the profiling buffer attempting to write to
memory are reported by the SPE hardware by asserting a buffer management
event interrupt. Interrupts are by their nature asynchronous, which means
that the guest might have changed its stage 1 translation tables since the
attempted write. SPE reports the guest virtual address that caused the data
abort, not the IPA, which means that KVM would have to walk the guest's
stage 1 tables to find the IPA. Using the AT instruction to walk the
guest's tables in hardware is not an option because it doesn't report the
IPA in the case of a stage 2 fault on a stage 1 table walk.

Avoid both issues by adding a capability that allows the user to pin the
memory backing a memslot and to pre-map it at stage 2. The same capability
can be used to unlock a memslot, which unpins the pages associated with the
memslot, but doesn't unmap the IPA range from stage 2.  The addresses will
be unmapped from stage 2 via the MMU notifiers when the process' address
space changes.

For now, the capability doesn't actually do anything other than checking
that the usage is correct; the memory operations will be added in future
patches.