KVM: arm64: Implement the memslot lock/unlock functionality
Pin memory in the process address space and map it in the stage 2 tables as
a result of userspace enabling the KVM_CAP_ARM_LOCK_USER_MEMORY_REGION
capability; and unpin it from the process address space when the capability
is used with the KVM_ARM_LOCK_USER_MEMORY_REGION_FLAGS_UNLOCK flag.

The current implementation has two drawbacks which will be fixed in future
patches:

- The dcache maintenance is done when the memslot is locked, which means
  that it is possible that memory changes made by userspace after the ioctl
  won't be visible to a guest running with the MMU off.

- The tag scrubbing is done when the memslot is locke. If the MTE
  capability is enabled after the ioctl, the guest will be able to access
  unsanitised pages. This is prevented by forbidding userspace to enable
  the MTE capability if any memslots are locked.

Only PAGE_SIZE mappings are supported at stage 2.