KVM: arm64: Save/restore PMSNEVFR_EL1 on VCPU put/load
SPEv1p3 introduced a new register, PMSNEVFR_EL1. The SPE driver is not
using the register, so save the register to the guest context on vcpu_put()
and restore it on vcpu_load() since it will not be touched by the host,
and the value programmed by the guest doesn't affect the host.